feat: 범용적으로 사용할 수 있도록 수정
